Split renderer/note-modal.tsx — extract icons, find, config-summary parts
